Moved err msgs, added header mod with write buf
the way I want this written means that the buffer used to write
should not fill in the first 8 bytes as they are reserved for a length
header. I do not expect the user to deal with this themselves because
that is tedious boilerplate. I'm just going to write a struct which
handles this for them with some basic methods.