Server support & QIO trait.
I realized that I had written myself into a corner on the server side
because the Qrexec struct which contained the write buffer only existed
on the client side; I rectified this situation by writing a struct for
the server side as well. This removed the need for genericism on the
read and write methods so read and write are now cleaner methods on the
structs for both sides of the vchan.